Move testing to unit tests
